+++ packages/utils/tsup.config.ts
@@ -1,5 +1,6 @@
 import { defineConfig } from 'tsup';
-import { readdirSync, statSync } from 'fs';
+import { readdirSync, statSync, writeFileSync, readFileSync } from 'fs';
+import { resolve } from 'path';
 
 export const allUtilsEntry = readdirSync('./src').reduce(
   (allUtils, util) => {
@@ -18,4 +19,27 @@ export const allUtilsEntry = readdirSync('./src').reduce(
 export default defineConfig({
   entry: allUtilsEntry,
   format: 'esm',
+  onSuccess: async () => {
+    const pkgPath = resolve('./dist/package.json');
+    const originalPkgPath = resolve('./package.json');
+    const originalPkg = JSON.parse(readFileSync(originalPkgPath, 'utf-8'));
+    
+    const exports = {
+      ".": {
+        "types": "./index.d.ts",
+        "import": "./index.js"
+      },
+      "./*": {
+        "types": "./*.d.ts",
+        "import": "./*.js"
+      }
+    };
+    
+    const distPkg = {
+      ...originalPkg,
+      exports
+    };
+    
+    writeFileSync(pkgPath, JSON.stringify(distPkg, null, 2));
+  }
 });